+++ NEWS.md
@@ -1,3 +1,8 @@
+## version 0.2.16
+### Features
+- update an existing JAGS fit with `JAGS_extend()` function
+- new element of the `autofit_control` argument in `JAGS_fit()`: `"restarts"` allows to restart model initialization up to `restarts` times in case of failure
+
 ## version 0.2.15
 ### Fixes
 - fixing repeated print of previous prior distribution in `model_summary_table()` in case of `prior_none()`